+++ lib/pact/mock_service/request_handlers.rb
@@ -1,4 +1,5 @@
 require 'pact/mock_service/request_handlers/interaction_post'
+require 'pact/mock_service/request_handlers/interactions_put'
 require 'pact/mock_service/request_handlers/index_get'
 require 'pact/mock_service/request_handlers/interaction_delete'
 require 'pact/mock_service/request_handlers/interaction_replay'
@@ -22,14 +23,15 @@ module Pact
         def initialize name, logger, session, options
           super [
             Options.new(name, logger, options[:cors_enabled]),
-            MissingInteractionsGet.new(name, logger, session.expected_interactions, session.actual_interactions),
-            VerificationGet.new(name, logger, session.expected_interactions, session.actual_interactions),
-            InteractionPost.new(name, logger, session.expected_interactions, session.verified_interactions),
-            InteractionDelete.new(name, logger, session.expected_interactions, session.actual_interactions),
+            MissingInteractionsGet.new(name, logger, session),
+            VerificationGet.new(name, logger, session),
+            InteractionPost.new(name, logger, session),
+            InteractionsPut.new(name, logger, session),
+            InteractionDelete.new(name, logger, session),
             LogGet.new(name, logger),
-            PactPost.new(name, logger, session.verified_interactions, session.consumer_contract_details),
+            PactPost.new(name, logger, session),
             IndexGet.new(name, logger),
-            InteractionReplay.new(name, logger, session.expected_interactions, session.actual_interactions, session.verified_interactions, options[:cors_enabled])
+            InteractionReplay.new(name, logger, session, options[:cors_enabled])
           ]
         end
       end