+++ lib/pact/mock_service/request_handlers/verification_get.rb
@@ -5,10 +5,10 @@ module Pact
     module RequestHandlers
       class VerificationGet < BaseAdministrationRequestHandler
 
-        def initialize name, logger, expected_interactions, actual_interactions
+        def initialize name, logger, session
           super name, logger
-          @expected_interactions = expected_interactions
-          @actual_interactions = actual_interactions
+          @expected_interactions = session.expected_interactions
+          @actual_interactions = session.actual_interactions
         end
 
         def request_path